Refactored trello to separate component
