+++ date-recommender/src/components/Home.vue
@@ -6,44 +6,20 @@
             <h1>Date Recommender</h1>
           </b-col>
         </b-row>
-        <b-row>
-          <b-col>
-            <b-button v-on:click="getRandomTrelloDate()" pill variant="outline-primary" class="mb-2">Existing planned date idea</b-button>
-            <b-card>
-              {{randomTrelloDate}}
-            </b-card>
-          </b-col>
-        </b-row>
+        <Trello />
         <Yelp />
       </div>
   </b-container>
 </template>
 
 <script>
-import TrelloService from './../services/TrelloService'
 import Yelp from './../components/Yelp.vue'
-
+import Trello from './../components/Trello.vue'
 export default {
   name: 'HelloWorld',
-  props: {
-    trelloResponse: Array,
-    randomTrelloDate: Object
-  },
-  methods: {
-    getRandomTrelloDate: function() {
-      this.randomTrelloDate = this.trelloResponse[Math.floor(Math.random()*this.trelloResponse.length)].name
-    }
-  },
   components: {
+    Trello,
     Yelp
-  },
-  mounted () {
-    TrelloService
-      .getCards()
-        .then(response => {
-          this.trelloResponse = response.data;
-          this.randomTrelloDate = response.data[0].name;
-        })
   }
 }
 </script>
